\name{OMERarr-package}
\alias{OMERarr-package}
\alias{OMERarr}
\docType{package}
\title{
\packageTitle{OMERarr}
}
\description{
\packageDescription{OMERarr}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{OMERarr}

Maintainer: \packageMaintainer{OMERarr}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = OMERarr)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{OMERarr}}
%\section{Documentation Index}{\packageIndices{OMERarr}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
