% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{OMEZarrArray-methods}
\alias{OMEZarrArray-methods}
\alias{[[,OMEZarrArray,numeric-method}
\alias{meta}
\alias{meta,ImageArray-method}
\alias{dim,OMEZarrArray-method}
\alias{type,OMEZarrArray-method}
\alias{length,OMEZarrArray-method}
\alias{meta,OMEZarrArray-method}
\alias{read_image}
\title{Methods for OMEZarrArray}
\usage{
\S4method{[[}{OMEZarrArray,numeric}(x, i)

\S4method{dim}{OMEZarrArray}(x)

\S4method{type}{OMEZarrArray}(x)

\S4method{length}{OMEZarrArray}(x)

\S4method{meta}{OMEZarrArray}(object)

read_image(group)
}
\arguments{
\item{x, object}{An OMEZarrArray object}

\item{i}{Depends on the usage
\describe{
 \item{\code{[[}}{
   Here \code{i} is the level of the image pyramid.
   You can use the \code{length} function to get the
   number of the layers in the pyramid.
 }
}}

\item{group}{description}

\item{s3_client}{Object created by \code{paws.storage::s3()}. 
Only required for a file on S3. Leave as NULL for a file on local storage.}
}
\value{
dim of the first level of the OMEZarrArray object

type of OMEZarrArray object

length of OMEZarrArray object

length of OMEZarrArray object

An OMEZarrArray object
}
\description{
Methods for \code{OMEZarrArray} objects
}
\section{Functions}{
\itemize{
\item \code{x[[i}: Layer access
for \code{OMEZarrArray} objects

\item \code{dim(OMEZarrArray)}: dimensions of an OMEZarrArray

\item \code{type(OMEZarrArray)}: dimensions of an OMEZarrArray

\item \code{length(OMEZarrArray)}: length of an OMEZarrArray

\item \code{meta(OMEZarrArray)}: length of an OMEZarrArray

\item \code{read_image()}: OMEZarrArray constructor method

A function for creating objects of OMEZarrArray class

}}
