% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_image}
\alias{write_image}
\title{write_image}
\usage{
write_image(
  image,
  group = "/",
  axes = NULL,
  scalefactor = 2,
  max_layer = 5,
  format = "v0.4",
  storage_options = NULL
)
}
\arguments{
\item{image}{an n-dimensional (1<n<6) array representing the image data}

\item{group}{the zarr group to write the image to}

\item{axes}{a character vector specifying the axes of the image 
(e.g. c("t", "c", "z", "y", "x"))}

\item{scalefactor}{Scaler implementation for downsampling the image argument. 
If None, no downsampling will be performed.}

\item{max_layer}{Maximum number of layers in the image pyramid}

\item{format}{OME-ZARR format (v0.4 or v0.5)}

\item{storage_options}{a list of storage options for the zarr array 
(e.g. chunks)}
}
\description{
Writes an image to the zarr store according to ome-zarr specification
}
